REFACTOR:Explore dummyData
